Release

+++ packages/backend/src/queue/processors/ImportCustomEmojisProcessorService.ts
@@ -84,9 +84,20 @@ export class ImportCustomEmojisProcessorService {
 					continue;
 				}
 				const emojiPath = outputPath + '/' + record.fileName;
-				await this.emojisRepository.delete({
-					name: emojiInfo.name,
-				});
+				const emojiExist = await this.emojisRepository.findOneBy({ name: emojiInfo.name });
+
+				if (emojiExist) {
+					await this.emojisRepository.delete({
+						name: emojiInfo.name,
+					});
+					const file = await this.driveFilesRepository.findOneBy({ id: emojiExist.driveId });
+
+					if (file == null) {
+						this.logger.error(`Cannot delete ${ emojiInfo.id } (${ emojiInfo.driveId }). Ignored.`);
+						continue;
+					}
+					await this.driveService.deleteFile(file);
+				}
 				try {
 					const driveFile = await this.driveService.addFile({
 						user: null,
@@ -95,15 +106,18 @@ export class ImportCustomEmojisProcessorService {
 						force: true,
 					});
 					await this.customEmojiService.add({
+						originalUrl: driveFile.url,
+						publicUrl: driveFile.webpublicUrl ?? driveFile.url,
+						fileType: driveFile.webpublicType ?? driveFile.type,
 						name: emojiInfo.name,
 						category: emojiInfo.category,
 						host: null,
 						aliases: emojiInfo.aliases,
-						driveFile,
 						license: emojiInfo.license,
 						isSensitive: emojiInfo.isSensitive,
 						localOnly: emojiInfo.localOnly,
 						roleIdsThatCanBeUsedThisEmojiAsReaction: [],
+						driveId: driveFile.id,
 					});
 				} catch (e) {
 					if (e instanceof Error || typeof e === 'string') {
